Account module would crash after using the Select Patient dropdown sometimes. (When using random primary keys)
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@5577 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Data Interface/PatientL.cs
@@ -67,7 +67,7 @@ namespace OpenDental{
 			int index=menu.MenuItems.IndexOf((MenuItem)sender);
 			//Patients.PatIsLoaded=true;
 			if(index<buttonLastFivePatNums.Count) {
-				return (int)buttonLastFivePatNums[index];
+				return (long)buttonLastFivePatNums[index];
 			}
 			if(fam==null) {
 				return 0;//will never happen
